split model creation and export for TEASER
