+++ bim2sim/plugins/PluginTEASER/bim2sim_teaser/__init__.py
@@ -8,6 +8,7 @@ from bim2sim.plugins.PluginTEASER.bim2sim_teaser.models import TEASER
 from bim2sim.tasks import common, bps, base
 from bim2sim.sim_settings import TEASERSimSettings
 
+
 class LoadLibrariesTEASER(base.ITask):
     """Load AixLib library for export"""
     touches = ('libraries', )
@@ -33,6 +34,7 @@ class PluginTEASER(Plugin):
         bps.CombineThermalZones,
         teaser_task.WeatherTEASER,
         LoadLibrariesTEASER,
+        teaser_task.CreateTEASER,
         teaser_task.ExportTEASER,
         teaser_task.SimulateModel,
     ]
