+++ bim2sim/plugins/PluginTEASER/bim2sim_teaser/task/__init__.py
@@ -1,3 +1,4 @@
 from .export_teaser import ExportTEASER
+from .create_teaser_prj import CreateTEASER
 from .simulate_dymola import SimulateModel
 from .weather import WeatherTEASER
